kafka: move protocol types out of common header
The kafka/types.h currently is expensive header. However, the protocol
library primarily operates on primitive types. It should not be required
to include large amount of expensive things that are never used.
