+++ src/v/kafka/protocol/kafka_batch_adapter.h
@@ -13,6 +13,7 @@
 
 #include "bytes/iobuf_parser.h"
 #include "kafka/types.h"
+#include "kafka/protocol/types.h"
 #include "model/record.h"
 #include "model/record_batch_reader.h"
 #include "storage/record_batch_builder.h"
