+++ src/v/kafka/types.h
@@ -12,6 +12,7 @@
 #pragma once
 #include "bytes/bytes.h"
 #include "bytes/details/out_of_range.h"
+#include "kafka/protocol/types.h"
 #include "model/fundamental.h"
 #include "reflection/adl.h"
 #include "seastarx.h"
@@ -27,42 +28,8 @@ namespace kafka {
 /// Kafka API request correlation.
 using correlation_id = named_type<int32_t, struct kafka_correlation_type>;
 
-/// Kafka API key.
-using api_key = named_type<int16_t, struct kafka_requests_api_key>;
-
-/// Kafka API version.
-using api_version = named_type<int16_t, struct kafka_requests_api_version>;
-
-/// Kafka group identifier.
-using group_id = named_type<ss::sstring, struct kafka_group_id>;
-
-/// Kafka transactional id identifier.
-using transactional_id = named_type<ss::sstring, struct kafka_transactional_id>;
-
-/// Kafka producer id identifier.
-using producer_id = named_type<int64_t, struct kafka_producer_id>;
-
-/// Kafka group member identifier.
-using member_id = named_type<ss::sstring, struct kafka_member_id>;
-
-/// Kafka group generation identifier.
-using generation_id = named_type<int32_t, struct kafka_generation_id>;
-
-/// Kafka group instance identifier.
-using group_instance_id
-  = named_type<ss::sstring, struct kafka_group_instance_id>;
-
-/// Kafka group protocol type.
-using protocol_type = named_type<ss::sstring, struct kafka_protocol_type>;
-
-/// Kafka group protocol name.
-using protocol_name = named_type<ss::sstring, struct kafka_protocol>;
-
 using client_id = named_type<ss::sstring, struct kafka_client_id_type>;
 using client_host = named_type<ss::sstring, struct kafka_client_host_type>;
-
-/// An unknown / missing member id (Kafka protocol specific)
-static inline const member_id unknown_member_id("");
 
 /// An unknown / missing generation id (Kafka protocol specific)
 static inline const generation_id unknown_generation_id(-1);
@@ -81,67 +48,6 @@ static constexpr fetch_session_epoch initial_fetch_session_epoch(0);
  */
 static constexpr fetch_session_epoch final_fetch_session_epoch(-1);
 
-enum class coordinator_type : int8_t {
-    group = 0,
-    transaction = 1,
-};
-
-using leader_epoch = named_type<int32_t, struct leader_epoch_tag>;
-/**
- * Used to mark that leader epoch is not intended to be used (Kafka protocol
- * specific)
- */
-static constexpr leader_epoch invalid_leader_epoch(-1);
-
-/**
- * Immutable UUID, represents a 128 bit value of the variant 2 (Leach-Salz)
- * version 4 UUID type. Conversions to/from string will expect or perform a b64
- * encoding/decoding
- */
-class uuid {
-public:
-    static constexpr auto length = 16;
-    using underlying_t = std::array<uint8_t, length>;
-
-    uuid() = default;
-
-    static uuid from_string(std::string_view encoded) {
-        if (encoded.size() > 24) {
-            details::throw_out_of_range(
-              "Input size of {} too long to be decoded as b64-UUID, expected "
-              "{} bytes or less",
-              encoded.size(),
-              24);
-        }
-        auto decoded = base64_to_bytes(encoded);
-        if (decoded.size() != length) {
-            details::throw_out_of_range(
-              "Expected {} byte value post b64decoding the input: {} bytes",
-              length,
-              decoded.size());
-        }
-        underlying_t ul;
-        std::copy_n(decoded.begin(), length, ul.begin());
-        return uuid(ul);
-    }
-
-    explicit uuid(const underlying_t& uuid)
-      : _uuid(uuid) {}
-
-    bytes_view view() const { return {_uuid.data(), _uuid.size()}; }
-
-    ss::sstring to_string() const { return bytes_to_base64(view()); }
-
-    friend bool operator==(const uuid&, const uuid&) = default;
-
-    friend std::ostream& operator<<(std::ostream& os, const uuid& u) {
-        return os << u.to_string();
-    }
-
-private:
-    underlying_t _uuid{};
-};
-
 // TODO Ben: Why is this an undefined reference for pandaproxy when defined in
 // kafka/requests.cc
 inline std::ostream& operator<<(std::ostream& os, coordinator_type t) {
@@ -153,12 +59,6 @@ inline std::ostream& operator<<(std::ostream& os, coordinator_type t) {
     };
     return os << "{unknown type}";
 }
-
-enum class config_resource_type : int8_t {
-    topic = 2,
-    broker = 4,
-    broker_logger = 8,
-};
 
 inline std::ostream& operator<<(std::ostream& os, config_resource_type t) {
     switch (t) {
@@ -180,22 +80,6 @@ enum class config_resource_operation : int8_t {
 };
 
 std::ostream& operator<<(std::ostream& os, config_resource_operation);
-
-/*
- * From where config values are sourced. For instance, a value might exist
- * because it is a default or because it was an override at the broker level.
- *
- * As our configuration becomes more sophisticated these should be taken into
- * account. Right now we only report a couple basic topic configs.
- */
-enum class describe_configs_source : int8_t {
-    topic = 1,
-    static_broker_config = 4,
-    default_config = 5,
-    // DYNAMIC_BROKER_CONFIG((byte) 2),
-    // DYNAMIC_DEFAULT_BROKER_CONFIG((byte) 3),
-    // DYNAMIC_BROKER_LOGGER_CONFIG((byte) 6);
-};
 
 inline std::ostream& operator<<(std::ostream& os, describe_configs_source s) {
     switch (s) {
